Move resolveProjectPathInfo into project-index module
