+++ src/plugin/src/identifier-case/local-plan.js
@@ -1,10 +1,8 @@
 import { formatIdentifierCase } from "./identifier-case-utils.js";
 import { buildRenameKey } from "./plan-state.js";
 import { asArray, isNonEmptyArray } from "../../../shared/array-utils.js";
-import {
-    toPosixPath,
-    resolveProjectPathInfo
-} from "../../../shared/path-utils.js";
+import { toPosixPath } from "../../../shared/path-utils.js";
+import { resolveProjectPathInfo } from "../project-index/path-info.js";
 import { createMetricsTracker } from "../../../shared/metrics.js";
 import {
     isNonEmptyString,
