+++ src/shared/path-utils.js
@@ -3,7 +3,6 @@ export {
     collectUniqueAncestorDirectories,
     fromPosixPath,
     isPathInside,
-    resolveProjectPathInfo,
     resolveContainedRelativePath,
     toPosixPath,
     walkAncestorDirectories
